+++ src/web/assets/ckeditor/CkeditorAsset.php
@@ -46,15 +46,17 @@ class CkeditorAsset extends BaseCkeditorPackageAsset
      * @inheritdoc
      */
     public $js = [
-        'ckeditor5-dll.js',
-        'ckeditor5-craftcms.js',
+        // Import map shim
+        ['https://ga.jspm.io/npm:es-module-shims@2.0.10/dist/es-module-shims.js', 'async' => true],
+        ['ckeditor5-craftcms.js', 'type' => 'module'],
     ];
 
     /**
      * @inheritdoc
      */
     public $css = [
-        'css/ckeditor5-craftcms.css',
+        'lib/ckeditor5.css',
+        'ckeditor.css',
     ];
 
     public function registerAssetFiles($view): void
@@ -75,6 +77,7 @@ class CkeditorAsset extends BaseCkeditorPackageAsset
                 'Site: {name}',
                 'This field doesn’t allow nested entries.',
             ]);
+
             $view->registerJsWithVars(fn($attach) => <<<JS
 Craft.showCkeditorInspector = $attach;
 JS, [
@@ -94,12 +97,8 @@ JS, [
             }
         }
 
-        $view->registerJsWithVars(
-            fn($refHandles) => <<<JS
-window.CKEditor5.craftcms.localizedRefHandles = $refHandles;
-JS,
-            [$refHandles],
-            View::POS_END,
-        );
+        $view->registerScriptWithVars(fn($refHandles) => <<<JS
+window.CKE_LOCALIZED_REF_HANDLES = $refHandles;
+JS, [$refHandles], View::POS_HEAD, ['type' => 'module']);
     }
 }