+++ src/web/assets/ckeditor/src/entries/entriesui.js
@@ -1,13 +1,13 @@
-import {Plugin} from 'ckeditor5/src/core';
 import {
   addListToDropdown,
   ButtonView,
+  Collection,
   createDropdown,
+  isWidget,
+  Plugin,
   ViewModel,
-} from 'ckeditor5/src/ui';
-import {Range} from 'ckeditor5/src/engine';
-import {Collection} from 'ckeditor5/src/utils';
-import {isWidget, WidgetToolbarRepository} from 'ckeditor5/src/widget';
+  WidgetToolbarRepository,
+} from 'ckeditor5';
 import {DoubleClickObserver} from '../observers/domevent';
 
 export default class CraftEntriesUI extends Plugin {